Split base Peras types into separate module
