+++ ouroboros-consensus/src/ouroboros-consensus/Ouroboros/Consensus/Peras/Voting/Committee.hs
@@ -20,7 +20,6 @@ import qualified Data.Map as Map
 import qualified Data.Map.NonEmpty as NEMap
 import Data.Maybe (isJust)
 import Data.Word (Word16, Word64)
-import Ouroboros.Consensus.Block.SupportsPeras (PerasSeatIndex (..))
 import qualified Ouroboros.Consensus.Committee.Class as Committee
 import Ouroboros.Consensus.Committee.Crypto (CryptoSupportsVRF (..))
 import Ouroboros.Consensus.Committee.EveryoneVotes
@@ -32,6 +31,7 @@ import Ouroboros.Consensus.Committee.WFA (SeatIndex (..))
 import Ouroboros.Consensus.Committee.WFALS (Cert (..), Vote (..), WFALS)
 import qualified Ouroboros.Consensus.Peras.Cert.V1 as V1
 import Ouroboros.Consensus.Peras.Crypto.BLS (PerasBLSCrypto)
+import Ouroboros.Consensus.Peras.Types (PerasSeatIndex (..))
 import qualified Ouroboros.Consensus.Peras.Vote.V1 as V1
 
 -- * Peras support for multiple voting committee implementations
