+++ ouroboros-consensus/src/ouroboros-consensus/Ouroboros/Consensus/Peras/Voting/Rules.hs
@@ -34,9 +34,7 @@ import Ouroboros.Consensus.Block.Abstract
   )
 import Ouroboros.Consensus.Block.SupportsPeras
   ( HasPerasCertRound (..)
-  , PerasRoundNo (..)
   , getPerasCertRound
-  , onPerasRoundNo
   )
 import Ouroboros.Consensus.Peras.Params
   ( PerasCertArrivalThreshold (..)
