+++ ouroboros-consensus/test/consensus-test/Test/Consensus/Peras/Util.hs
@@ -37,17 +37,17 @@ import Data.Word (Word8)
 import GHC.Word (Word16)
 import Ouroboros.Consensus.Block (HeaderHash)
 import Ouroboros.Consensus.Block.RealPoint (RealPoint (..), toBytes32RealPoint)
-import Ouroboros.Consensus.Block.SupportsPeras
-  ( PerasBoostedBlock (..)
-  , PerasRoundNo (..)
-  , PerasSeatIndex (..)
-  )
 import qualified Ouroboros.Consensus.Committee.Crypto.BLS as BLS
 import qualified Ouroboros.Consensus.Peras.Cert.V1 as V1
 import Ouroboros.Consensus.Peras.Crypto.BLS
   ( PerasBLSCryptoAggregateVoteSignature (..)
   , VRFOutput (..)
   , VoteSignature (..)
+  )
+import Ouroboros.Consensus.Peras.Types
+  ( PerasBoostedBlock (..)
+  , PerasRoundNo (..)
+  , PerasSeatIndex (..)
   )
 import qualified Ouroboros.Consensus.Peras.Vote.V1 as V1
 import Test.QuickCheck
