+++ ouroboros-consensus/test/consensus-test/Test/Consensus/Peras/Voting/Rules.hs
@@ -19,9 +19,7 @@ import Ouroboros.Consensus.Block.Abstract
   )
 import Ouroboros.Consensus.Block.SupportsPeras
   ( HasPerasCertRound (..)
-  , PerasRoundNo (..)
   , getPerasCertRound
-  , onPerasRoundNo
   )
 import Ouroboros.Consensus.BlockchainTime
   ( RelativeTime (..)
