+++ ouroboros-consensus/test/storage-test/Test/Ouroboros/Storage/PerasVoteDB/Model.hs
@@ -27,22 +27,29 @@ import Ouroboros.Consensus.Block.SupportsPeras
   , HasPerasVoteRound (..)
   , PerasCert (..)
   , PerasCfg
-  , PerasParams (..)
-  , PerasRoundNo
-  , PerasVoteId (..)
-  , PerasVoteStake (..)
-  , PerasVoteTarget (..)
-  , PerasVoterId
   , ValidatedPerasCert (..)
   , ValidatedPerasVote
   , getPerasCertBoostedBlock
   , getPerasVoteStake
   , getPerasVoteVoterId
-  , stakeAboveThreshold
   )
 import Ouroboros.Consensus.BlockchainTime.WallClock.Types
   ( WithArrivalTime (..)
   )
+import Ouroboros.Consensus.Peras.Params (PerasParams, perasWeight)
+import Ouroboros.Consensus.Peras.Types
+  ( PerasRoundNo
+  , PerasVoteId (..)
+  , PerasVoteStake (..)
+  , PerasVoteTarget (..)
+  , PerasVoterId
+  , ValidatedPerasCert (..)
+  , ValidatedPerasVote (..)
+  , perasWeight
+  , stakeAboveThreshold
+  )
+import Ouroboros.Consensus.BlockchainTime.WallClock.Types (WithArrivalTime (..))
+import Ouroboros.Consensus.Peras.Cert.Mock (MockPerasCert (..))
 import Ouroboros.Consensus.Storage.PerasVoteDB.API
   ( AddPerasVoteResult (..)
   , PerasVoteTicketNo
